requirements.txt
attrs==21.4.0
click==8.0.4
flasgger==0.9.5
Flask==2.0.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
mistune==2.0.2
psycopg2==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
PyYAML==6.0
six==1.16.0
Werkzeug==2.0.3
zipp==3.7.0
